# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Resume Agent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/reviews/page": {
      "post": {
        "summary": "Start Page",
        "operationId": "start_page_api_reviews_page_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews/flow": {
      "post": {
        "summary": "Start Flow",
        "operationId": "start_flow_api_reviews_flow_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlowReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews/{run_id}/result": {
      "get": {
        "summary": "Result",
        "operationId": "result_api_reviews__run_id__result_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews/{run_id}/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_api_reviews__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews/{run_id}/cancel": {
      "post": {
        "summary": "Cancel",
        "operationId": "cancel_api_reviews__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews/{run_id}/retry": {
      "post": {
        "summary": "Retry",
        "operationId": "retry_api_reviews__run_id__retry_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews/{run_id}/record": {
      "get": {
        "summary": "Record",
        "operationId": "record_api_reviews__run_id__record_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews/retry": {
      "post": {
        "summary": "Retry Stored",
        "operationId": "retry_stored_api_reviews_retry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoredRetryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews/{run_id}": {
      "delete": {
        "summary": "Release",
        "operationId": "release_api_reviews__run_id__delete",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CancelAccepted": {
        "properties": {
          "runId": {
            "type": "string",
            "title": "Runid"
          },
          "cancelRequested": {
            "type": "boolean",
            "title": "Cancelrequested"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "runId",
          "cancelRequested"
        ],
        "title": "CancelAccepted"
      },
...
      "FlowDocument": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/FlowBlock"
            },
            "type": "array",
            "minItems": 1,
            "title": "Blocks"
          },
          "extraction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionReport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "blocks"
        ],
        "title": "FlowDocument"
      },
...
      "FlowReviewRequest": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/FlowDocument"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/ReferenceMaterial"
            },
            "type": "array",
            "title": "Materials"
          },
          "previousReview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviousReview"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviewContext": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewContext"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document"
        ],
        "title": "FlowReviewRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageDocument": {
        "properties": {
          "pages": {
            "items": {
              "$ref": "#/components/schemas/DocumentPage"
            },
            "type": "array",
            "minItems": 1,
            "title": "Pages"
          },
          "extraction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionReport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pages"
        ],
        "title": "PageDocument"
      },
...
      "PageReviewRequest": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/PageDocument"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/ReferenceMaterial"
            },
            "type": "array",
            "title": "Materials"
          },
          "previousReview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviousReview"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviewContext": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewContext"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document"
        ],
        "title": "PageReviewRequest"
      },
      "PreviousReview": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SuggestionDTO"
            },
            "type": "array",
            "title": "Results"
          },
          "userFeedback": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Userfeedback"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "results"
        ],
        "title": "PreviousReview"
      },
      "ReferenceMaterial": {
        "properties": {
          "sourceId": {
            "type": "string",
            "minLength": 1,
            "title": "Sourceid"
          },
          "materialType": {
            "$ref": "#/components/schemas/MaterialType"
          },
          "inputType": {
            "type": "string",
            "enum": [
              "text",
              "document",
              "url"
            ],
            "title": "Inputtype"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sourceId",
          "materialType",
          "inputType"
        ],
        "title": "ReferenceMaterial"
      },
      "ResolutionCheck": {
        "properties": {
          "suggestionId": {
            "type": "string",
            "minLength": 1,
            "title": "Suggestionid"
          },
          "status": {
            "type": "string",
            "enum": [
              "resolved",
              "notApplied",
              "uncertain"
            ],
            "title": "Status"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SuggestionLocation"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "suggestionId",
          "status",
          "reason"
        ],
        "title": "ResolutionCheck"
      },
...
      "RetryRequest": {
        "properties": {
          "moduleKey": {
            "type": "string",
            "enum": [
              "spellCheck",
              "companyContextAnalysis",
              "jobPostingAnalysis",
              "finalReview"
            ],
            "title": "Modulekey"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "moduleKey"
        ],
        "title": "RetryRequest"
      },
...
      "ReviewContext": {
        "properties": {
          "contextId": {
            "type": "string",
            "minLength": 1,
            "title": "Contextid"
          },
          "resumeVersionId": {
            "type": "string",
            "minLength": 1,
            "title": "Resumeversionid"
          },
          "materialVersions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Materialversions"
          },
          "previousReviewId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previousreviewid"
          },
          "feedback": {
            "items": {
              "$ref": "#/components/schemas/SuggestionFeedback"
            },
            "type": "array",
            "title": "Feedback"
          },
          "resolved": {
            "items": {
              "$ref": "#/components/schemas/ResolvedMemory"
            },
            "type": "array",
            "title": "Resolved"
          },
          "userFeedback": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Userfeedback"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/ExperienceCandidate"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Experiences"
          },
          "selectedExperienceIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Selectedexperienceids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contextId",
          "resumeVersionId"
        ],
        "title": "ReviewContext"
      },
...
      "ReviewRecord": {
        "properties": {
          "runId": {
            "type": "string",
            "minLength": 1,
            "title": "Runid"
          },
          "reviewContext": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageDocument"
              },
              {
                "$ref": "#/components/schemas/FlowDocument"
              }
            ],
            "title": "Document"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/ReferenceMaterial"
            },
            "type": "array",
            "title": "Materials"
          },
          "previousReview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviousReview"
              },
              {
                "type": "null"
              }
            ]
          },
          "moduleResults": {
            "$ref": "#/components/schemas/CompletedModules"
          },
          "response": {
            "$ref": "#/components/schemas/ReviewResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "runId",
          "createdAt",
          "document",
          "materials",
          "moduleResults",
          "response"
        ],
        "title": "ReviewRecord",
        "description": "JSON record shared by storage and retry; contains no runtime objects."
      },
      "ReviewResponse": {
        "properties": {
          "experienceRecommendations": {
            "items": {
              "$ref": "#/components/schemas/ExperienceRecommendation"
            },
            "type": "array",
            "title": "Experiencerecommendations"
          },
          "resolutionChecks": {
            "items": {
              "$ref": "#/components/schemas/ResolutionCheck"
            },
            "type": "array",
            "title": "Resolutionchecks"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ModuleErrorDTO"
            },
            "type": "array",
            "title": "Errors"
          },
          "materialReviews": {
            "items": {
              "$ref": "#/components/schemas/MaterialReviewDTO"
            },
            "type": "array",
            "title": "Materialreviews"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SuggestionDTO"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "errors",
          "materialReviews",
          "results"
        ],
        "title": "ReviewResponse"
      },
...
      "RunAccepted": {
        "properties": {
          "runId": {
            "type": "string",
            "title": "Runid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "runId"
        ],
        "title": "RunAccepted"
      },
...
      "StoredRetryRequest": {
        "properties": {
          "moduleKey": {
            "type": "string",
            "enum": [
              "spellCheck",
              "companyContextAnalysis",
              "jobPostingAnalysis",
              "finalReview"
            ],
            "title": "Modulekey"
          },
          "record": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviewId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "moduleKey"
        ],
        "title": "StoredRetryRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}